Implement new prettify_obs class
