+++ src/cell_annotator/__init__.py
@@ -5,7 +5,16 @@ from .check import check_deps
 from .model._api_keys import APIKeyManager
 from .model.base_annotator import BaseAnnotator
 from .model.cell_annotator import CellAnnotator
+from .model.obs_beautifier import ObsBeautifier
 from .model.sample_annotator import SampleAnnotator
 
-__all__ = ["CellAnnotator", "SampleAnnotator", "BaseAnnotator", "logger", "check_deps", "APIKeyManager"]
+__all__ = [
+    "CellAnnotator",
+    "SampleAnnotator",
+    "BaseAnnotator",
+    "logger",
+    "check_deps",
+    "APIKeyManager",
+    "ObsBeautifier",
+]
 __version__ = version("cell-annotator")
