+++ src/cell_annotator/model/cell_annotator.py
@@ -7,17 +7,10 @@ from tqdm.auto import tqdm
 from cell_annotator._constants import PackageConstants
 from cell_annotator._docs import d
 from cell_annotator._logging import logger
-from cell_annotator._response_formats import CellTypeColorOutput, CellTypeListOutput, ExpectedMarkerGeneOutput
+from cell_annotator._response_formats import CellTypeListOutput, ExpectedMarkerGeneOutput
 from cell_annotator.model.base_annotator import BaseAnnotator
 from cell_annotator.model.sample_annotator import SampleAnnotator
-from cell_annotator.utils import (
-    _filter_marker_genes_to_adata,
-    _format_annotation,
-    _get_consistent_ordering,
-    _get_unique_cell_types,
-    _try_sorting_dict_by_keys,
-    _validate_list_mapping,
-)
+from cell_annotator.utils import _filter_marker_genes_to_adata, _format_annotation, _try_sorting_dict_by_keys
 
 
 @d.dedent
@@ -359,139 +352,3 @@ class CellAnnotator(BaseAnnotator):
         logger.info("Iterating over samples to harmonize cell type annotations.")
         for annotator in tqdm(self.sample_annotators.values()):
             annotator.harmonize_annotations(self.global_cell_type_list, unknown_key=unknown_key)
-
-    def reorder_and_color_clusters(
-        self, keys: list[str] | str, unknown_key: str = PackageConstants.unknown_name, assign_colors: bool = False
-    ) -> None:
-        """Assign consistent ordering across cell type annotations.
-
-        Note that for multiple samples with many clusters each, this typically requires a more powerful model
-        to work well. This method replaces underscores with spaces.
-
-        Parameters
-        ----------
-        keys
-            List of keys in `adata.obs` to reorder.
-        unknown_key
-            Name of the unknown category.
-        assign_colors
-            Assign colors to the cell types across keys. These are supposed to be consistent across keys and meaningful bioligically,
-            such that similar cell types get similar colors.
-
-        Returns
-        -------
-        Updates the following attributes:
-        - `self.adata.obs[keys]`
-        """
-        if isinstance(keys, str):
-            keys = [keys]
-
-        # make the naming consistent: replaces underscores with spaces
-        for key in keys:
-            self.adata.obs[key] = self.adata.obs[key].map(lambda x: x.replace("_", " "))
-
-        # Take the union of all cell types across keys and re-order the list
-        cell_type_list = self._get_cluster_ordering(keys, unknown_key=unknown_key)
-
-        # assign meaningful colors to the reordered list of cell types
-        if assign_colors:
-            global_names_and_colors = self._get_cluster_colors(clusters=cell_type_list, unknown_key=unknown_key)
-        else:
-            global_names_and_colors = dict.fromkeys(cell_type_list, "")
-
-        label_sets = _get_consistent_ordering(self.adata, global_names_and_colors, keys)
-
-        # Re-add the unknown category, make sure that we retained all categories, and write to adata
-        for obs_key, name_and_color in label_sets.items():
-            if unknown_key in self.adata.obs[obs_key].cat.categories:
-                logger.debug(
-                    "Readding the unknown category with key '%s'",
-                    unknown_key,
-                )
-                name_and_color[unknown_key] = PackageConstants.unknown_color
-
-            _validate_list_mapping(list(self.adata.obs[obs_key].unique()), list(name_and_color.keys()), context=obs_key)
-
-            logger.info("Writing categories for key '%s'", obs_key)
-            self.adata.obs[obs_key] = self.adata.obs[obs_key].cat.set_categories(list(name_and_color.keys()))
-            if assign_colors:
-                self.adata.uns[f"{obs_key}_colors"] = name_and_color.values()
-
-    def _get_cluster_ordering(self, keys: list[str], unknown_key: str = PackageConstants.unknown_name) -> list[str]:
-        """Query LLM for relational cluster ordering.
-
-        Parameters
-        ----------
-        keys
-            List of keys in `adata.obs` whose categories should be ordered.
-        unknown_key
-            Name of the unknown category.
-
-        Returns
-        -------
-        List of cell types in some biologically meaningful order.
-        """
-        # format the current annotations sets as a string and prepare the query prompt
-        unique_cell_types = _get_unique_cell_types(self.adata, keys, unknown_key)
-        order_prompt = self.prompts.get_order_prompt(unique_cell_types=", ".join(unique_cell_types))
-
-        # query llm and format the response as a dict
-        logger.info("Querying label ordering.")
-        response = self.query_llm(
-            instruction=order_prompt,
-            response_format=CellTypeListOutput,
-        )
-
-        return response.cell_type_list
-
-    def _get_cluster_colors(
-        self, clusters: str | list[str], unknown_key: str = PackageConstants.unknown_name
-    ) -> dict[str, str]:
-        """Query LLM for relational cluster colors.
-
-        Parameters
-        ----------
-        clusters
-            Either a key in `adata.obs` or a list of cell type names.
-        unknown_key
-            Name of the unknown category.
-
-        Returns
-        -------
-        Mapping of cell types to colors.
-
-        """
-        if isinstance(clusters, str):
-            if clusters not in self.adata.obs:
-                raise ValueError(f"Key '{clusters}' not found in `adata.obs`.")
-            cluster_list = list(self.adata.obs[clusters].unique())
-        elif isinstance(clusters, list):
-            cluster_list = clusters
-        else:
-            raise ValueError(f"Invalid type for 'clusters': {type(clusters)}")
-
-        cluster_names = ", ".join(cl for cl in cluster_list if cl != unknown_key)
-
-        logger.info("Querying cluster colors.")
-        response = self.query_llm(
-            instruction=self.prompts.get_color_prompt(cluster_names), response_format=CellTypeColorOutput
-        )
-        color_dict = {
-            item.original_cell_type_label: item.assigned_color for item in response.cell_type_to_color_mapping
-        }
-
-        # Re-add the unknown category, make sure that we retained all categories, and write to adata
-        if unknown_key in cluster_list:
-            logger.debug(
-                "Readding the unknown category with key '%s'",
-                unknown_key,
-            )
-            color_dict[unknown_key] = PackageConstants.unknown_color
-
-        # make sure that the new categories are the same as the original categories
-        _validate_list_mapping(cluster_list, list(color_dict.keys()))
-
-        # fix their order in case it was changed
-        color_dict = {key: color_dict[key] for key in cluster_list}
-
-        return color_dict
